Rename fs -> sys to avoid conflict with builtins
